doc(logs): correct format argument guidance (#2027)
Use the standard %d argument type in the example and remove the stale
guidance requiring 64-bit arguments.

Ref: #1752

+++ examples/example.c
@@ -1036,8 +1036,7 @@ main(int argc, char **argv)
         sentry_value_set_by_key(attributes, "number.first", attr_2);
         sentry_value_set_by_key(attributes, "number.second", attr_3);
         // testing multiple attributes
-        sentry_log_debug(
-            "logging with %d custom attributes", attributes, (int64_t)3);
+        sentry_log_debug("logging with %d custom attributes", attributes, 3);
         // testing no attributes
         sentry_log_debug("logging with %s custom attributes",
             sentry_value_new_object(), "no");
